+++ src/Commands/img/glitch.js
@@ -20,7 +20,7 @@ module.exports = class GlitchCommand extends Command {
 					key: 'image',
 					prompt: 'What image would you like to edit?',
 					type: 'image',
-					default: (msg) => msg.author.displayAvatarURL({ format: 'png', size: 4096 }),
+					default: (msg) => msg.author.displayAvatarURL({ format: 'png', size: 512 }),
 				},
 			],
 		});