+++ src/Commands/img/triggered.js
@@ -21,7 +21,7 @@ module.exports = class TriggeredCommand extends Command {
 					key: 'image',
 					prompt: 'What image would you like to edit?',
 					type: 'image',
-					default: (msg) => msg.author.displayAvatarURL({ format: 'png', size: 4096 }),
+					default: (msg) => msg.author.displayAvatarURL({ format: 'png', size: 512 }),
 				},
 			],
 		});